my-ml-mcp-server/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
streamlit>=1.28.0
pyyaml>=6.0
pydantic>=2.0
requests>=2.28.0
